Rename io packages to iolocal
Rename packages as "io" is a common first segment of package names (io.circe for example) and can hence be very annoying to users that then have to sort their imports or prefix them with _root_.
