--- modules/io/src/main/scala/io/janstenpickle/trace4cats/base/context/io/instances.scala
@@ -1,3 +1,0 @@
-package io.janstenpickle.trace4cats.base.context.io
-
-object instances extends IOLocalContextInstances
