+++ modules/io/src/main/scala/io/janstenpickle/trace4cats/base/context/iolocal/IOLocalContextInstances.scala
@@ -1,4 +1,4 @@
-package io.janstenpickle.trace4cats.base.context.io
+package io.janstenpickle.trace4cats.base.context.iolocal
 
 import cats.effect.{IO, IOLocal}
 import io.janstenpickle.trace4cats.base.context.Provide
