+++ modules/io/src/main/scala/io/janstenpickle/trace4cats/base/context/iolocal/instances.scala
@@ -1,0 +1,3 @@
+package io.janstenpickle.trace4cats.base.context.iolocal
+
+object instances extends IOLocalContextInstances
