+++ modules/io/src/main/scala/io/janstenpickle/trace4cats/inject/iolocal/package.scala
@@ -1,0 +1,5 @@
+package io.janstenpickle.trace4cats.inject
+
+import io.janstenpickle.trace4cats.base.context.iolocal.IOLocalContextInstances
+
+package object iolocal extends IOLocalTraceInstances with IOLocalContextInstances
